+++ packages/adminui-vue/components/UTableEntity/filter-templates/date/notIsNull.vue
@@ -1,0 +1,24 @@
+<template>
+  <filter-template @submit="submitHandler" />
+</template>
+
+<script>
+  export default {
+    name: 'FilterDateNotIsNull',
+
+    components: {
+      FilterTemplate: require('../../components/FilterTemplate.vue').default
+    },
+
+    mixins: [require('../mixinForFilter.js')],
+
+    methods: {
+      getCondition () {
+        return {
+          whereList: [{ condition: 'notIsNull' }],
+          description: 'notIsNull'
+        }
+      }
+    }
+  }
+</script>