+++ packages/adminui-vue/components/UTableEntity/filter-templates/string/notIsNull.vue
@@ -1,0 +1,26 @@
+<template>
+  <filter-template @submit="submitHandler"/>
+</template>
+
+<script>
+export default {
+  name: 'FilterStringNotIsNull',
+  components: {
+    FilterTemplate: require('../../components/FilterTemplate.vue').default
+  },
+
+  mixins: [require('../mixinForFilter.js')],
+
+  created () {
+    this.$emit('search-disabled', false)
+  },
+  methods: {
+    getCondition () {
+      return {
+        whereList: [{ condition: 'notIsNull' }],
+        description: 'notIsNull'
+      }
+    }
+  }
+}
+</script>